Moved middleware tests to a separate file.
